+++ tests/trace/test_file_storage_credentials.py
@@ -3,49 +3,11 @@ from unittest import mock
 
 import pytest
 
-from weave.trace_server.file_management import (
+from weave.trace_server.file_storage_credentials import (
     get_aws_credentials,
     get_azure_credentials,
     get_gcp_credentials,
-    parse_storage_uri,
-    split_bucket_and_path,
 )
-
-
-def test_parse_storage_uri():
-    # Test valid URIs
-    assert parse_storage_uri("s3://bucket/path") == ("s3", "bucket/path")
-    assert parse_storage_uri("gs://bucket/path") == ("gs", "bucket/path")
-    assert parse_storage_uri("azure://container/path") == ("azure", "container/path")
-
-    # Test invalid URIs
-    with pytest.raises(ValueError, match="Invalid storage URI format"):
-        parse_storage_uri("")
-    with pytest.raises(ValueError, match="Invalid storage URI format"):
-        parse_storage_uri("invalid-uri")
-    with pytest.raises(ValueError, match="No path specified"):
-        parse_storage_uri("s3://")
-    with pytest.raises(ValueError, match="Unsupported storage provider"):
-        parse_storage_uri("invalid://bucket/path")
-
-
-def test_split_bucket_and_path():
-    # Test valid paths
-    assert split_bucket_and_path("bucket/path", "s3") == ("bucket", "path")
-    assert split_bucket_and_path("bucket/path/nested", "gs") == (
-        "bucket",
-        "path/nested",
-    )
-    assert split_bucket_and_path("container/blob/path", "azure") == (
-        "container",
-        "blob/path",
-    )
-
-    # Test invalid paths
-    with pytest.raises(ValueError, match="Invalid path format"):
-        split_bucket_and_path("invalid-path", "s3")
-    with pytest.raises(ValueError, match="Invalid path format"):
-        split_bucket_and_path("", "gs")
 
 
 def test_get_aws_credentials():
@@ -53,9 +15,9 @@ def test_get_aws_credentials():
     with mock.patch.dict(
         os.environ,
         {
-            "WF_STORAGE_BUCKET_AWS_ACCESS_KEY_ID": "test-key",
-            "WF_STORAGE_BUCKET_AWS_SECRET_ACCESS_KEY": "test-secret",
-            "WF_STORAGE_BUCKET_AWS_SESSION_TOKEN": "test-token",
+            "WF_FILE_STORAGE_BUCKET_AWS_ACCESS_KEY_ID": "test-key",
+            "WF_FILE_STORAGE_BUCKET_AWS_SECRET_ACCESS_KEY": "test-secret",
+            "WF_FILE_STORAGE_BUCKET_AWS_SESSION_TOKEN": "test-token",
         },
     ):
         creds = get_aws_credentials()
@@ -68,8 +30,8 @@ def test_get_aws_credentials():
     with mock.patch.dict(
         os.environ,
         {
-            "WF_STORAGE_BUCKET_AWS_ACCESS_KEY_ID": "test-key",
-            "WF_STORAGE_BUCKET_AWS_SECRET_ACCESS_KEY": "test-secret",
+            "WF_FILE_STORAGE_BUCKET_AWS_ACCESS_KEY_ID": "test-key",
+            "WF_FILE_STORAGE_BUCKET_AWS_SECRET_ACCESS_KEY": "test-secret",
         },
     ):
         creds = get_aws_credentials()
@@ -89,7 +51,7 @@ def test_get_azure_credentials():
     with mock.patch.dict(
         os.environ,
         {
-            "WF_STORAGE_BUCKET_AZURE_CONNECTION_STRING": "test-connection-string",
+            "WF_FILE_STORAGE_BUCKET_AZURE_CONNECTION_STRING": "test-connection-string",
         },
     ):
         creds = get_azure_credentials()
@@ -102,8 +64,8 @@ def test_get_azure_credentials():
     with mock.patch.dict(
         os.environ,
         {
-            "WF_STORAGE_BUCKET_AZURE_ACCOUNT_URL": "test-url",
-            "WF_STORAGE_BUCKET_AZURE_CREDENTIAL": "test-credential",
+            "WF_FILE_STORAGE_BUCKET_AZURE_ACCOUNT_URL": "test-url",
+            "WF_FILE_STORAGE_BUCKET_AZURE_CREDENTIAL": "test-credential",
         },
     ):
         creds = get_azure_credentials()
@@ -136,7 +98,7 @@ def test_get_gcp_credentials():
     with mock.patch.dict(
         os.environ,
         {
-            "WF_STORAGE_BUCKET_GCP_CREDENTIALS_JSON": test_creds_json,
+            "WF_FILE_STORAGE_BUCKET_GCP_CREDENTIALS_JSON": test_creds_json,
         },
     ):
         with mock.patch("google.oauth2.service_account.Credentials") as mock_creds:
@@ -152,7 +114,7 @@ def test_get_gcp_credentials():
     with mock.patch.dict(
         os.environ,
         {
-            "WF_STORAGE_BUCKET_GCP_CREDENTIALS_JSON": "invalid-json",
+            "WF_FILE_STORAGE_BUCKET_GCP_CREDENTIALS_JSON": "invalid-json",
         },
     ):
         with pytest.raises(ValueError, match="Invalid GCP credentials JSON"):
